Переменные и некоторые функции перенесены в отдельные файлы, привел код в порядок
